pyproject.toml
[project]
name = "eodo"
version = "0.1.9"
description = "EdgeOneDynamicOrigin"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "PyYAML",
    "streamlit",
    "schedule",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
eodo = "eodo.app:main"

[tool.hatch.build]
sources = ["src"]
